// src/entities/sidebar/types/ETabRoutes.ts
export enum ETabRoutes {
  HOME = '/',
  ELIGIBILITY_QUIZ = '/eligibility-quiz',
  BASIC_PERSONAL_INFO = '/basic-personal-info',
  FAMILY_AND_MARRIAGE = '/family-and-marriage',
  IMMIGRATION_AND_TRAVEL = '/immigration-and-travel',
  EMPLOYMENT = '/employment',
  FINANCIALS = '/financials',
  DISCLAIMER = '/disclaimer',
  MISCELLANEOUS = '/miscellaneous',
  ADMIN_PANEL = '/admin-panel',
  FORMS = '/forms',
}